+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Fixed duplicated active worktime entries by trying to fix it automatically.
 * Fixed clocking in when multiple worktimes are active with QRclock plugin.
+* Added `CustomRoutes::getCustomRoutes()` and `CustomRoutes::getCustomRoute($route)` functions to get all custom routes/one specific route file.
 
 <!-- Fixed missing API route for Codeclock plugin. -->